(#25): 추상화
